rp2040: Avoid run-time division in hard_pwm.c
Use a MAX_PWM of 256 (instead of 255) to avoid an expensive run-time
divide operation.

+++ src/rp2040/gpio.h
@@ -26,7 +26,7 @@ struct gpio_pwm {
     uint8_t shift;
     uint32_t mask;
 };
-struct gpio_pwm gpio_pwm_setup(uint8_t pin, uint32_t cycle_time, uint8_t val);
+struct gpio_pwm gpio_pwm_setup(uint8_t pin, uint32_t cycle_time, uint32_t val);
 void gpio_pwm_write(struct gpio_pwm g, uint32_t val);
 
 struct gpio_adc {
